requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
dnspython==1.16.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycparser==2.20
pylint==2.6.0
pymongo==3.11.2
PyNaCl==1.4.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.12.7